fix(react): share composer input headless state
